Fix broken backpressure by combining `HTTPServerRequestDecoder` and `HTTPServerResponseEncoder` into a single `ChannelDuplexHandler`
